proper response and exception handling are implemented

+++ src/main/java/com/nayemuf/employee_management/service/EmployeeService.java
@@ -1,7 +1,9 @@
 package com.nayemuf.employee_management.service;
 
 import com.nayemuf.employee_management.entity.Employee;
+import com.nayemuf.employee_management.exception.ResourceNotFoundException;
 import com.nayemuf.employee_management.repository.EmployeeRepository;
+import com.nayemuf.employee_management.response.ApiResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -21,9 +23,12 @@ public class EmployeeService {
         return (List<Employee>) employeeRepository.findAll();
     }
 
-    public Optional<Employee> getEmployeeById(int id) {
-        return employeeRepository.findById(id);
+    public Employee getEmployeeById(int id) {
+        // Fetch employee by ID
+        return employeeRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Employee not found with ID: " + id));
     }
+
 
 
     public Employee updateEmployee(int id, Employee updatedEmployee) {
@@ -33,11 +38,16 @@ public class EmployeeService {
             employee.setDepartment(updatedEmployee.getDepartment());
             employee.setSalary(updatedEmployee.getSalary());
             return employeeRepository.save(employee);
-        }).orElse(null);
+        }).orElseThrow(() -> new ResourceNotFoundException("Employee not found with id: " + id));
     }
 
-    public String deleteEmployeeById(int id){
+    public ApiResponse<String> deleteEmployeeById(int id){
+
+        if (!employeeRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Employee not found with ID: " + id);
+        }
+
         employeeRepository.deleteById(id);
-        return "Employee deleted successfully";
+        return new ApiResponse<String>(true, "Employee successfully deleted", null);
     }
 }
